feat(war): implement pagination in findAll method

+++ src/wars/wars.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { CreateWarDto } from './dto/create-war.dto';
 import { UpdateWarDto } from './dto/update-war.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -6,17 +6,22 @@ import { Repository } from 'typeorm';
 import { War } from './entities/war.entity';
 import { plainToInstance } from 'class-transformer';
 import { UUID } from 'crypto';
+import {
+  DEFAULT_TAKE_VALUE,
+  PaginationOptions,
+} from 'src/helpers/pagination/interface';
 
 @Injectable()
 export class WarsService {
+  private readonly logger = new Logger(WarsService.name);
+
   constructor(
     @InjectRepository(War)
     private warRepository: Repository<War>,
   ) {}
 
   async create(createWarDto: CreateWarDto) {
-    console.log('createWarDto', createWarDto);
-    const war = plainToInstance(War, createWarDto);
+    const war = await this.warRepository.create(createWarDto);
     return await this.warRepository.save(war);
   }
 
@@ -29,11 +34,30 @@ export class WarsService {
     } else throw new NotFoundException();
   }
 
-  findAll({ take = 10, skip = 0 }: { take?: number; skip: number }) {
-    const limit = take > 100 ? 100 : (take ?? 10);
-    const offset = skip > 1000 ? 1000 : (skip ?? 0);
+  findAll(options: PaginationOptions): Promise<War[]> {
+    //before should default to the current date
+    const {
+      limit = DEFAULT_TAKE_VALUE,
+      after,
+      before = new Date(),
+      dec = false,
+    } = options;
+    const forceLimit = Math.min(limit, 10);
 
-    return this.warRepository.findAndCount({ take: limit, skip: offset });
+    const query = this.warRepository.createQueryBuilder('war');
+
+    if (after) {
+      query.andWhere('war.startAt > :after', { after: new Date(after) });
+    }
+
+    if (before) {
+      query.andWhere('war.startAt < :before', { before: new Date(before) });
+    }
+
+    query.orderBy('war.startAt', dec ? 'DESC' : 'ASC');
+    query.take(forceLimit);
+
+    return query.getMany();
   }
 
   findOne(id: UUID) {
@@ -41,7 +65,9 @@ export class WarsService {
   }
 
   async remove(id: UUID) {
-    this.warRepository.delete(id);
+    const entity = await this.warRepository.findOneBy({ id });
+    if (!entity) throw new NotFoundException(`War with ID ${id} not found`);
+    await this.warRepository.delete(id);
   }
   // async findAllInWar(warId: UUID, leagueId?: UUID) {
   //   const participants = await this.participantRepository.find({
